% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_standardize.R
\name{aggregate_pop}
\alias{aggregate_pop}
\title{Aggregate population arrays}
\usage{
aggregate_pop(pop, margin, groups = NULL, bind_new = FALSE, new_name = NULL)
}
\arguments{
\item{pop}{The population array to aggregate}

\item{margin}{The margin on which the groups of interest are stratified}

\item{groups}{A vector of either indices for each group or a vector of strings for each group name. If set to \code{NULL}, will use all groups in the dataset.}

\item{bind_new}{If set to \code{TRUE}, will bind to the original sample dataset using \code{abind::abind()}. Otherwise, will generate a standalone array of samples.}

\item{new_name}{If binding the array to the original sample dataset, will assign a name to the new margin}
}
\value{
An \code{array} of aggregated population data
}
\description{
Aggregate population arrays
}
\examples{
margin_time = 3
# aggregate population from all years for each county-group
pop_7988 = aggregate_pop(miheart$n, margin_time)
# aggregate population from 1980-1984 for each county-group
pop_8084 = aggregate_pop(miheart$n, margin_time, groups = as.character(1980:1984))
# bind aggregated pop from all years to population data
pop_agg = aggregate_pop(miheart$n, margin_time, bind_new = TRUE, new_name = "1979-1988")
}
