% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstr_statops.R
\name{anova_vec}
\alias{anova_vec}
\title{A vectorized version of analysis of variance (ANOVA).}
\usage{
anova_vec(rstr_lm_full, rstr_lm_null, rstr_data)
}
\arguments{
\item{rstr_lm_full}{An object of type \code{\link{RstrModel}} returned from \code{\link{lm_vec}}.
This is a full model including both the main effect and covariates.}

\item{rstr_lm_null}{An object of type \code{\link{RstrModel}} returned from \code{\link{lm_vec}}.
This is a null model including only the covariates.}

\item{rstr_data}{Object of type \code{\link{RstrData}}}
}
\description{
This function compares results of model fitting after \code{\link{lm_vec}}.
It accepts a full model and and a reduced model and compares them using an F-test.
For most scenarios, the user does not need to call this function directly. This function
will be called internally from \code{\link{rstr_anova}}
}
\seealso{
\code{\link{rstr_anova}} for most commonly used function for ANOVA, \code{\link{lm_vec}} for vectorized linear regression, \code{\link{ttest_vec}} for
vectorized independent sample and paired t-tests.
}
