% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstroutput_class.R
\name{check_files}
\alias{check_files}
\title{Check that subject directory and demographics files exist}
\usage{
check_files(object)
}
\arguments{
\item{object}{object of type \code{RstrOutput}}
}
\description{
Check that subject directory and demographics files exist
}
