% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstr_utils.R
\name{check_multiple_files_exists}
\alias{check_multiple_files_exists}
\title{Check if multiple files exists all at once}
\usage{
check_multiple_files_exists(filelist, errmesg = NULL)
}
\arguments{
\item{filelist}{Name of file.}

\item{errmesg}{character string of optional error message}
}
\description{
Check if multiple files exists all at once
}
