% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstr_utils.R
\name{coord2ind}
\alias{coord2ind}
\title{Convert a 3D coordinate from an image to a linear vector index
Equivalent to sub2ind from MATLAB(R)
Does not perform error checking for bounds}
\usage{
coord2ind(coord, dimI)
}
\arguments{
\item{coord}{vector of 3D coordinates}

\item{dimI}{vector of image dimensions}
}
\description{
Convert a 3D coordinate from an image to a linear vector index
Equivalent to sub2ind from MATLAB(R)
Does not perform error checking for bounds
}
