% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstrdata.R
\name{copy_sba_data}
\alias{copy_sba_data}
\title{Package cortical surface data for reproducible statistical analysis.}
\usage{
copy_sba_data(
  subjdir = "",
  csv = "",
  hemi = "left",
  smooth = 0,
  outdir,
  exclude_col = ""
)
}
\arguments{
\item{subjdir}{subject directory containing MouseSuite processed data.}

\item{csv}{filename of a comma separated (csv) file containing the subject demographic information.
The first column of this csv file
should be "subjID" and should have subject identifiers you wish to analyze. subjID can be alphanumeric
and should be exactly equal to the individual subject directory name.}

\item{hemi}{chaaracter string denoting the brain hemisphere. Should either be "left" or "right".}

\item{smooth}{numeric value denoting the smoothing level.}

\item{outdir}{output directory that will contain the copied data}

\item{exclude_col}{character string for the column in demographics csv (contains 1 or 0 for each row) specifying the subjects to exclude. 1 denotes include, 0 denotes exclude.}
}
\description{
Package cortical surface data for reproducible statistical analysis.
}
