% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_model_auxiliary.R
\name{generate_data}
\alias{generate_data}
\title{\code{generate_data()} converts a long \code{table} featuring event counts across regions and other optional margins into a \code{list} that is readable by \code{*car()}.
\code{generate_data()} will sum along any stratifications that aren't used. Filter data by desired groups and time periods before running \code{generate_data()}.}
\usage{
generate_data(table, event, population, region, group = NULL, time = NULL)
}
\arguments{
\item{table}{A \code{table} containing event and mortality counts}

\item{event}{The column containing event counts}

\item{population}{The column containing population counts}

\item{region}{The column containing region names}

\item{group}{An optional column containing sociodemographic group names}

\item{time}{An optional column containing time period names}
}
\value{
A \code{list} of mortality and population counts organized into multi-dimensional arrays
}
\description{
\code{generate_data()} converts a long \code{table} featuring event counts across regions and other optional margins into a \code{list} that is readable by \code{*car()}.
\code{generate_data()} will sum along any stratifications that aren't used. Filter data by desired groups and time periods before running \code{generate_data()}.
}
\examples{
ma_data <- maexample[!is.na(maexample$Year), ]
ma_data_mst <- generate_data(maexample, Deaths, Population, County.Code, Sex.Code, Year.Code) # Generates data from 1979-1981 stratified by sex
ma_data_79 <- ma_data[ma_data$Year == 1979, ]
ma_data_m <- generate_data(ma_data_79, Deaths, Population, County.Code, Sex.Code) # Generates 1979 data stratified by sex
ma_data_u <- generate_data(ma_data_79, Deaths, Population, County.Code) # Generates 1979 data summarized for all sexes
}
