% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstr_setup.R
\name{get_brainsuite_path_from_rstr_ini}
\alias{get_brainsuite_path_from_rstr_ini}
\title{Retrieve BrainSuite installation path from rstr.ini}
\usage{
get_brainsuite_path_from_rstr_ini()
}
\description{
Retrieve BrainSuite installation path from rstr.ini
}
