% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medians.R
\name{get_credible_interval}
\alias{get_credible_interval}
\title{Calculate Credible Intervals}
\usage{
get_credible_interval(output, perc_ci = 0.95)
}
\arguments{
\item{output}{Array of output to generate relative precisions for.}

\item{perc_ci}{Number from 0 to 1. Determines width of credible interval.}
}
\description{
Calculate Credible Intervals
}
