% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_file_formats.R
\name{get_custom_tbm_atlas_and_mask}
\alias{get_custom_tbm_atlas_and_mask}
\title{Get the custom tensor based morphometry atlas and mask}
\usage{
get_custom_tbm_atlas_and_mask(mousesuite_custom_atlas_prefix)
}
\arguments{
\item{mousesuite_custom_atlas_prefix}{file prefix for atlas}
}
\description{
Get the custom tensor based morphometry atlas and mask
}
