% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstr_colormaps.R
\name{get_logpvalue_colormap}
\alias{get_logpvalue_colormap}
\title{Get log pvalue colormap}
\usage{
get_logpvalue_colormap(cmap_name, values)
}
\arguments{
\item{cmap_name}{name of the colormap}

\item{values}{log-transformed p-values}
}
\description{
Get log pvalue colormap
}
