% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_file_formats.R
\name{get_mousesuite_atlas_path_from_logfile}
\alias{get_mousesuite_atlas_path_from_logfile}
\title{Read the MouseSuite atlas path from the svreg log file.}
\usage{
get_mousesuite_atlas_path_from_logfile(logfile)
}
\arguments{
\item{logfile}{path to the log file present in an individual subject directory}
}
\description{
Read the MouseSuite atlas path from the svreg log file.
}
