% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_file_formats.R
\name{get_mousesuite_custom_volume_atlas_prefix}
\alias{get_mousesuite_custom_volume_atlas_prefix}
\title{Read the MouseSuite atlas prefix from the atlas}
\usage{
get_mousesuite_custom_volume_atlas_prefix(atlas)
}
\arguments{
\item{atlas}{filepath for atlas}
}
\description{
Read the MouseSuite atlas prefix from the atlas
}
