% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_file_formats.R
\name{get_mousesuite_logfilename_for_all_subjects}
\alias{get_mousesuite_logfilename_for_all_subjects}
\title{Get the svreg log file for each subject}
\usage{
get_mousesuite_logfilename_for_all_subjects(subjdir, csv, exclude_col = "")
}
\arguments{
\item{subjdir}{individual subject directory that the svreg.log file exists in}

\item{csv}{csv file for the svreg.log file}

\item{exclude_col}{character string for the column in demographics csv (contains 1 or 0 for each row) specifying the subjects to exclude. 1 denotes include, 0 denotes exclude.}
}
\description{
Get the svreg log file for each subject
}
