% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_file_formats.R
\name{get_render_image_filename}
\alias{get_render_image_filename}
\title{Returns the filename for the designated image}
\usage{
get_render_image_filename(
  outdir,
  voxelcoord,
  overlay_name,
  brain_sector_index,
  voxelcoord_index
)
}
\arguments{
\item{outdir}{string denoting the output directory}

\item{voxelcoord}{all outputted voxelcoordinates}

\item{overlay_name}{string denoting the name of the overlay}

\item{brain_sector_index}{numeric value denoting which brain slice is desired}

\item{voxelcoord_index}{numeric value denoting desired voxel coordinate}
}
\description{
Returns the filename for the designated image
}
