% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_file_formats.R
\name{get_roi_file_list}
\alias{get_roi_file_list}
\title{Returns a list of all ROI files for all subjects}
\usage{
get_roi_file_list(rstr_data, roistat_fileprefix, roilabeldescfile)
}
\arguments{
\item{rstr_data}{object of type \code{RstrData}}

\item{roistat_fileprefix}{file prefix for the ROI stats file (either .tissue.stats.txt or .roiwise.stats.txt)",}

\item{roilabeldescfile}{an xml file containing ROI label descriptions",}
}
\description{
Returns a list of all ROI files for all subjects
}
