% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeldesc_io.R
\name{get_roi_tag}
\alias{get_roi_tag}
\title{Get the ROI tag from the ROI label}
\usage{
get_roi_tag(label_desc_df, roiid)
}
\arguments{
\item{label_desc_df}{\code{\link{data.frame}} containing fields from the label description file}

\item{roiid}{ROI label identifier}
}
\description{
Get the ROI tag from the ROI label
}
