% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_file_formats.R
\name{get_sba_atlas}
\alias{get_sba_atlas}
\title{Get the desired cortical surface atlas}
\usage{
get_sba_atlas(mousesuite_atlas_id, hemi)
}
\arguments{
\item{mousesuite_atlas_id}{individual subject directory that the log file exists in}

\item{hemi}{designates which hemisphere of the brain}
}
\description{
Get the desired cortical surface atlas
}
