% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_file_formats.R
\name{get_sba_file_list}
\alias{get_sba_file_list}
\title{Returns a list of the cortical surface files for all subjects}
\usage{
get_sba_file_list(rstr_data, hemi, smooth = 0)
}
\arguments{
\item{rstr_data}{object of type \code{RstrData}}

\item{hemi}{designates which hemisphere is of interest}

\item{smooth}{numeric value designating the smoothing used (default is 0)}
}
\description{
Returns a list of the cortical surface files for all subjects
}
