% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_file_formats.R
\name{get_tbm_atlas}
\alias{get_tbm_atlas}
\title{Get the MouseSuite tensor based morphometry atlas}
\usage{
get_tbm_atlas(mousesuite_atlas_id)
}
\arguments{
\item{mousesuite_atlas_id}{individual subject directory that the log file exists in}
}
\description{
Get the MouseSuite tensor based morphometry atlas
}
