% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_file_formats.R
\name{get_tbm_atlas_and_mask}
\alias{get_tbm_atlas_and_mask}
\title{Get the tensor based morphometry atlas and mask}
\usage{
get_tbm_atlas_and_mask(mousesuite_atlas_id)
}
\arguments{
\item{mousesuite_atlas_id}{individual subject directory that the svreg.log file exists in}
}
\description{
Get the tensor based morphometry atlas and mask
}
