% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_file_formats.R
\name{get_tbm_file_list}
\alias{get_tbm_file_list}
\title{Returns a list of the tensor-based files for all subjects}
\usage{
get_tbm_file_list(rstr_data, smooth = 0)
}
\arguments{
\item{rstr_data}{object of type \code{RstrData}}

\item{smooth}{numeric value designating the smoothing used (default is 0)}
}
\description{
Returns a list of the tensor-based files for all subjects
}
