% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_file_formats.R
\name{get_tbm_mask}
\alias{get_tbm_mask}
\title{Get the MouseSuite tensor based morphometry mask}
\usage{
get_tbm_mask(mousesuite_atlas_id)
}
\arguments{
\item{mousesuite_atlas_id}{individual subject directory that the svreg.log file exists in}
}
\description{
Get the MouseSuite tensor based morphometry mask
}
