% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstroutput_class.R
\name{get_voxelcoord}
\alias{get_voxelcoord}
\title{Get voxel coordinates of all significant clusters (up to number of clusters)}
\usage{
get_voxelcoord(rstr_out, rstr_data, rstr_model, outdir, nclusters)
}
\arguments{
\item{rstr_out}{object of type \code{RstrOut}}

\item{rstr_data}{object of type \code{RstrData}}

\item{rstr_model}{object of type \code{RstrModel}}

\item{outdir}{string specifying output directory to save the results in}

\item{nclusters}{numeric value specifying number of clusters}
}
\description{
Get voxel coordinates of all significant clusters (up to number of clusters)
}
