% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstr_utils.R
\name{ind2coord}
\alias{ind2coord}
\title{Convert a linear array index to a 3D coordinate in an image
Equivalent to ind2sub from MATLAB(R)
Does not perform error checking for bounds}
\usage{
ind2coord(idx, dimI)
}
\arguments{
\item{idx}{numeric index in a linear vector}

\item{dimI}{vector of image dimensions}
}
\description{
Convert a linear array index to a 3D coordinate in an image
Equivalent to ind2sub from MATLAB(R)
Does not perform error checking for bounds
}
