% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_model.R
\name{initialize_model}
\alias{initialize_model}
\title{Initialize CAR model}
\usage{
initialize_model(
  name,
  dir = tempdir(),
  data,
  adjacency,
  inits = NULL,
  priors = NULL,
  model = c("mstcar", "ucar", "mcar"),
  method = c("binomial", "poisson"),
  m0 = NULL,
  A = NULL,
  rho_up = FALSE,
  impute_lb = 1,
  impute_ub = 9,
  seed = 1234,
  show_plots = TRUE,
  ignore_checks = FALSE
)
}
\arguments{
\item{name}{Name of model and corresponding folder}

\item{dir}{Directory where model will live}

\item{data}{Dataset including mortality (Y) and population (n) information}

\item{adjacency}{Dataset including adjacency information}

\item{inits}{Optional list of initial conditions for each parameter}

\item{priors}{Optional list of priors for updates}

\item{model}{Run model as an MSTCAR/UCAR/MCAR model}

\item{method}{Run model with either Binomial data or Poisson data}

\item{m0}{Baseline neighbor count by region}

\item{A}{Describes intensity of smoothing between regions}

\item{rho_up}{Controls whether rho update is performed for MSTCAR models}

\item{impute_lb}{If counts are suppressed for privacy reasons, \code{impute_lb} is lower bound of suppression, typically 0 or 1}

\item{impute_ub}{If counts are suppressed for privacy reasons, \code{impute_ub} is upper bound of suppression, typically 10}

\item{seed}{Set of random seeds to use for data replication}

\item{show_plots}{If set to \code{FALSE}, suppresses check plots generated for MSTCAR models}

\item{ignore_checks}{If set to \code{TRUE}, ignores data checks. Only use if you are certain that your input data is
correct and you are encountering bugs during setup}
}
\value{
No output, only sets up model and saves files to directory
}
\description{
This function performs checks and prepares data for use with either an MSTCAR, MCAR, or UCAR model. This function additionally specifies all of the model parameters, such as model type, event data type, intensity of smoothing in the UCAR model, and more.
}
\examples{
# Initialize an MSTCAR model
initialize_model("test", dir = tempdir(), data = miheart, adjacency = miadj)
# Initialize an MCAR model
data_m <- lapply(miheart, \(x) x[, , "1979"])
initialize_model("test", tempdir(), data_m, miadj, model = "mcar")
# Initialize an MCAR model with Poisson-distributed event data
initialize_model("test", tempdir(), data_m, miadj, model = "mcar", method = "poisson")
# Initialize a restricted UCAR model
data_u <- lapply(miheart, \(x) x[, "65-74", "1979"])
initialize_model("test", tempdir(), data_u, miadj, model = "ucar", A = 6)
\dontshow{
unlink(paste0(tempdir(), "\\\\test"), recursive = TRUE)
}
}
