% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_model.R
\name{initialize_ucar}
\alias{initialize_ucar}
\alias{initialize_ucar_restricted}
\alias{initialize_mcar}
\alias{initialize_mstcar}
\title{Initialize CAR model}
\usage{
initialize_ucar(
  name,
  data,
  adjacency,
  dir = tempdir(),
  show_plots = TRUE,
  ignore_checks = FALSE,
  method = c("binomial", "poisson"),
  impute_lb = 1,
  impute_ub = 9,
  seed = 1234,
  inits = NULL,
  priors = NULL
)

initialize_ucar_restricted(
  name,
  data,
  adjacency,
  dir = tempdir(),
  A = NULL,
  m0 = NULL,
  show_plots = TRUE,
  ignore_checks = FALSE,
  method = c("binomial", "poisson"),
  impute_lb = 1,
  impute_ub = 9,
  seed = 1234,
  inits = NULL,
  priors = NULL
)

initialize_mcar(
  name,
  data,
  adjacency,
  dir = tempdir(),
  show_plots = TRUE,
  ignore_checks = FALSE,
  method = c("binomial", "poisson"),
  impute_lb = 1,
  impute_ub = 9,
  seed = 1234,
  inits = NULL,
  priors = NULL
)

initialize_mstcar(
  name,
  data,
  adjacency,
  dir = tempdir(),
  show_plots = TRUE,
  ignore_checks = FALSE,
  method = c("binomial", "poisson"),
  impute_lb = 1,
  impute_ub = 9,
  seed = 1234,
  inits = NULL,
  priors = NULL,
  rho_up = FALSE
)
}
\arguments{
\item{name}{Name of model and corresponding folder}

\item{data}{Dataset including mortality (Y) and population (n) information}

\item{adjacency}{Dataset including adjacency information}

\item{dir}{Directory where model will live}

\item{show_plots}{If set to \code{FALSE}, suppresses check plots generated for MSTCAR models}

\item{ignore_checks}{If set to \code{TRUE}, ignores data checks. Only use if you are certain that your input data is
correct and you are encountering bugs during setup}

\item{method}{Run model with either Binomial data or Poisson data}

\item{impute_lb}{If counts are suppressed for privacy reasons, \code{impute_lb} is lower bound of suppression, typically 0 or 1}

\item{impute_ub}{If counts are suppressed for privacy reasons, \code{impute_ub} is upper bound of suppression, typically 10}

\item{seed}{Set of random seeds to use for data replication}

\item{inits}{Optional list of initial conditions for each parameter}

\item{priors}{Optional list of priors for updates}

\item{A}{For restricted UCAR models, describes intensity of smoothing between regions}

\item{m0}{For restricted UCAR models, baseline neighbor count by region}

\item{rho_up}{For MSTCAR models, controls whether rho update is performed for MSTCAR models}
}
\value{
No output, only sets up model and saves files to directory
}
\description{
This function performs checks and prepares data for use with either an MSTCAR, MCAR, or UCAR model. This function additionally specifies all of the model parameters, such as model type, event data type, intensity of smoothing in the UCAR model, and more.
}
\examples{
# Initialize an MSTCAR model
initialize_mstcar(name = "test", data = miheart, adjacency = miadj, dir = tempdir())
# Initialize an MCAR model
data_m <- lapply(miheart, \(x) x[, , "1979"])
initialize_mcar("test", data_m, miadj, tempdir())
# Initialize an MCAR model with Poisson-distributed event data
initialize_mcar("test", data_m, miadj, tempdir(), method = "poisson")
# Initialize a restricted UCAR model
data_u <- lapply(miheart, \(x) x[, "65-74", "1979"])
initialize_ucar_restricted("test", data_u, miadj, tempdir(), A = 6)
\dontshow{
unlink(paste0(tempdir(), "\\\\test"), recursive = TRUE)
}
}
