% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstr_setup.R
\name{is_brainsute_installed}
\alias{is_brainsute_installed}
\title{Check if BrainSuite is installed.}
\usage{
is_brainsute_installed(quiet = FALSE, raise_error = FALSE)
}
\arguments{
\item{quiet}{boolean specifying whether warnings/messages should be displayed}

\item{raise_error}{boolean specifying whether an exception should be raised}
}
\description{
Check if the BrainSuite installation is valid by verifying if the appropriate
atlas files and data exist. This function is called from \code{\link{.onLoad}}
when the package is loaded. It opens \code{rstr.ini} and checks if all the
paths are valid.
}
