% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstr_statops.R
\name{lmer_vox}
\alias{lmer_vox}
\title{This function performs voxelwise ANOVA Linear Mixed Effects analysis using lmer}
\usage{
lmer_vox(
  group_var,
  main_effect = "",
  time_effect = "",
  covariates = "",
  rstr_data
)
}
\arguments{
\item{group_var}{Categorical variable name. This should be present in the demographics csv file associated
with \code{rstr_data}.}

\item{main_effect}{Character string containing an independent variable whose effect you want to measure.
It could be disease status, age, gender etc. This should strictly be a single variable. This can be
either a categorical or a continuous variable.}

\item{time_effect}{Character string denoting the time effect variable}

\item{covariates}{Character string containing a set of other predictors (variables) in the model. If more than
one covariates are included, they should be separated by a \code{+} operator similar to an R formula.}

\item{rstr_data}{Object of type \code{\link{RstrData}}}
}
\description{
This function performs voxelwise ANOVA Linear Mixed Effects analysis using lmer
}
