% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstrdata.R
\name{load_data}
\alias{load_data}
\alias{load_data,RstrData-method}
\alias{load_data,RstrSBAData-method}
\alias{load_data,RstrTBMData-method}
\alias{load_data,RstrDBAData-method}
\alias{load_data,RstrROIData-method}
\title{A generic function to load data for statistical analysis.}
\usage{
load_data(
  rstr_data,
  atlas_filename = NULL,
  maskfile = NULL,
  hemi = "left",
  measure = "",
  smooth = 0,
  eddy = TRUE,
  roiids = NULL,
  roimeas = NULL,
  exclude_col,
  roistat_fileprefix,
  roilabeldescfile
)

\S4method{load_data}{RstrData}(
  rstr_data,
  roiids = NULL,
  roimeas = NULL,
  exclude_col,
  roistat_fileprefix,
  roilabeldescfile
)

\S4method{load_data}{RstrSBAData}(rstr_data, atlas_filename, hemi, smooth)

\S4method{load_data}{RstrTBMData}(rstr_data, atlas_filename, maskfile = NULL, smooth)

\S4method{load_data}{RstrDBAData}(rstr_data, atlas_filename, maskfile = NULL, measure, smooth, eddy)

\S4method{load_data}{RstrROIData}(
  rstr_data,
  roiids = NULL,
  roimeas = NULL,
  exclude_col,
  roistat_fileprefix,
  roilabeldescfile
)
}
\arguments{
\item{rstr_data}{object of type \code{\link{RstrData}}}

\item{atlas_filename}{path name to the atlas}

\item{maskfile}{path name to the mask file}

\item{hemi}{chaaracter string denoting the brain hemisphere. Should either be "left" or "right" or "both".}

\item{measure}{character specifying the brain imaging measure. If analyzing diffusion data, should be "FA".}

\item{smooth}{numeric value denoting the smoothing level.}

\item{eddy}{boolean for specifying if the diffusion images were eddy-current corrected or not.}

\item{roiids}{numeric label identifier for the region of interest (ROI) type analysis.}

\item{roimeas}{character string for the ROI measure.}

\item{exclude_col}{character string for the column in demographics csv (contains 1 or 0 for each row) specifying the subjects to exclude. 1 denotes include, 0 denotes exclude.}

\item{roistat_fileprefix}{file prefix for the ROI stats file (either .tissue.stats.txt or .roiwise.stats.txt)",}

\item{roilabeldescfile}{an xml file containing ROI label descriptions",}
}
\description{
A generic function to load data for statistical analysis.
}
\details{
For the most part, the user will never have to call this function directly.
Instead the user should call \code{\link{load_rstr_data}}.
}
\seealso{
\code{\link{load_rstr_data}}
}
