% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstrdata.R
\name{load_roi_data}
\alias{load_roi_data}
\title{Load ROI data for statistical analysis.}
\usage{
load_roi_data(
  subjdir = "",
  csv = "",
  roiids = "",
  roimeas = "",
  exclude_col,
  roistat_fileprefix,
  roilabeldescfile
)
}
\arguments{
\item{subjdir}{subject directory containing MouseSuite processed data.}

\item{csv}{filename of a comma separated (csv) file containing the subject demographic information.
The first column of this csv file
should be "subjID" and should have subject identifiers you wish to analyze. subjID can be alphanumeric
and should be exactly equal to the individual subject directory name.}

\item{roiids}{numeric label identifiers for the regions of interest (ROI) type analysis.}

\item{roimeas}{character string for the ROI measure.}

\item{exclude_col}{character string for the column in demographics csv (contains 1 or 0 for each row) specifying the subjects to exclude. 1 denotes include, 0 denotes exclude.}

\item{roistat_fileprefix}{file prefix for the ROI stats file (either .tissue.stats.txt or .roiwise.stats.txt)",}

\item{roilabeldescfile}{an xml file containing ROI label descriptions",}
}
\description{
Load ROI data for statistical analysis.
}
