% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstrdata.R
\name{load_rstr_data_from_filelist}
\alias{load_rstr_data_from_filelist}
\title{Load data for statistical analysis from a filelist}
\usage{
load_rstr_data_from_filelist(
  csv = "",
  subjdir = "",
  hemi = "left",
  type = "sba",
  file_col = "",
  atlas = "",
  maskfile = ""
)
}
\arguments{
\item{csv}{filename of a comma separated (csv) file containing the subject demographic information.
The first column of this csv file
should be "subjID" and should have subject identifiers you wish to analyze. subjID can be alphanumeric
and should be exactly equal to the individual subject directory name.}

\item{subjdir}{subject directory containing MouseSuite processed data.}

\item{hemi}{chaaracter string denoting the brain hemisphere. Should either be "left" or "right".}

\item{type}{character string denoting type of analysis. Should be sba, tbm, or roi.}

\item{file_col}{character string for the full file path.}

\item{atlas}{character specifying the file path prefix (all characters in the file name upto the first ".") for the custom atlas. If empty, the atlas will be read from the svreg.log file in the subject directory.
Otherwise, for example, if the atlas for tensor based morphometry is located at /path/to/atlas/myatlas.mri.bfc.nii.gz, then specify atlas="/path/to/atlas/myatlas".}

\item{maskfile}{optional filename of the mask for tbm or diffusion parameter analysis. The mask has to be in the atlas space.}
}
\description{
Loading data is usually the first step before running any statistical analysis.
Prior to using this function, MouseSuite and svreg should be run on all subjects.
If required, smoothing should be performed on cortical surface or volumetric image based measures.
Unlike \code{\link{load_rstr_data}}, this function loads data from a csv that contains a column for filelist
A csv file containing subject demographic information should exist. The first column of this csv file
should have the subject identifiers. Subject identifiers can be alphanumeric
and should be exactly equal to the individual subject directory names.
This csv file should also contain a column that contains a full path to the data file to be loaded.
}
\examples{
\dontrun{
my_data <- load_rstr_data_from_filelist(csv = "/path/to/my/demographics.csv",
type="sba", file_col = "COL_NAME", atlast = "/path/to/atlas",
maskfile = "/path/to/maskfile")
}

}
