% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstr_mathops.R
\name{log10_transform}
\alias{log10_transform}
\title{log10 transform}
\usage{
log10_transform(values)
}
\arguments{
\item{values}{numeric vector}
}
\value{
numeric vector containing the log10 transformed \code{values}
}
\description{
log10 transform
}
