% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstrdata.R
\name{mask_low_var_voxels}
\alias{mask_low_var_voxels}
\title{Mask near constant voxels or voxels with low variance}
\usage{
mask_low_var_voxels(rstr_data)
}
\arguments{
\item{rstr_data}{RstrData}
}
\description{
Mask near constant voxels or voxels with low variance
}
