% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minsample}
\alias{minsample}
\title{Samples Generated for Michigan data}
\format{
`minsample`
a small sample dataset to be used in examples, generated by an MSTCAR model with two regions, three age groups, and three time periods.
}
\usage{
minsample
}
\description{
A 4-dimensional \code{array} of samples generated by the MSTCAR Gibbs sampler for use in testing and examples
}
\keyword{datasets}
