% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minsplit}
\alias{minsplit}
\title{Age- and Sex-stratified Samples for Michigan data}
\format{
`minsplit`
a small sample dataset to be used for demonstration with \code{split_sample_groups()}. Generated by an MSTCAR model with two regions, six age-sex groups, and three time periods.
}
\usage{
minsplit
}
\description{
A 4-dimensional \code{array} of samples generated by the MSTCAR Gibbs sampler for demonstration with \code{split_sample_groups()}
}
\keyword{datasets}
