% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstrmodel.R
\name{parse_lm}
\alias{parse_lm}
\title{Returns an error if necessary elements are missing in the model}
\usage{
parse_lm(
  main_effect = "",
  covariates = "",
  corr_var = "",
  group_var = "",
  model_type = "",
  demographics
)
}
\arguments{
\item{main_effect}{string designating which column of demographics is the main effect}

\item{covariates}{string designating which column of demographics is the covariates}

\item{corr_var}{string designating which column of demographics is the correlation variable}

\item{group_var}{string designating which column of demographics is the group variable}

\item{model_type}{string designating the type of model}

\item{demographics}{data frame of the demographics}
}
\description{
Returns an error if necessary elements are missing in the model
}
