% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeldesc_io.R
\name{read_label_desc}
\alias{read_label_desc}
\title{Read MouseSuite label description file}
\usage{
read_label_desc(label_desc_filename)
}
\arguments{
\item{label_desc_filename}{path to the label description file}
}
\description{
Read MouseSuite label description file
}
