% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi_io.R
\name{read_roi_data_for_all_subjects}
\alias{read_roi_data_for_all_subjects}
\title{Reads the MouseSuite ROI data file saved for each subject and returns the filenames}
\usage{
read_roi_data_for_all_subjects(
  roi_filelist,
  demographics,
  roiids,
  roimeas,
  roistat_fileprefix,
  roilabeldescfile
)
}
\arguments{
\item{roi_filelist}{filenames for the ROIwise stats in the individual subject directory}

\item{demographics}{data frame containing the demographics of all subjects}

\item{roiids}{vector of numeric label identifiers for the region of interest (ROI) type analysis.}

\item{roimeas}{character string for the ROI measure.}

\item{roistat_fileprefix}{file prefix for the ROI stats file (either .tissue.stats.txt or .roiwise.stats.txt)",}

\item{roilabeldescfile}{an xml file containing ROI label descriptions",}
}
\description{
Reads the MouseSuite ROI data file saved for each subject and returns the filenames
}
