% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfsio.R
\name{readdfsattributes}
\alias{readdfsattributes}
\title{Read dfs file attributes only}
\usage{
readdfsattributes(filename)
}
\arguments{
\item{filename}{filename of the dfs surface file.}
}
\value{
numeric vector of length \eqn{T}, where \eqn{T} is the number of attributes
(same as number of vertices) stored in the dfs file.
}
\description{
Reads the dfs file attributes (\url{http://brainsuite.org/formats/dfs/}).
}
