% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendervols.R
\name{render_statmap_on_atlas}
\alias{render_statmap_on_atlas}
\title{Render statistical map overlays on the rodent atlas}
\usage{
render_statmap_on_atlas(
  outdir,
  rstr_data,
  rstr_model,
  var_name,
  stats_string,
  voxelcoord,
  alpha = 120
)
}
\arguments{
\item{outdir}{path for the output directory to save the results}

\item{rstr_data}{object of type `RstrData`}

\item{rstr_model}{object of type `RstrModel`}

\item{var_name}{string specifying the statistical variable}

\item{stats_string}{list of strings specifying the statistical measure}

\item{voxelcoord}{list of voxel coordinates (peak cluster value)}

\item{alpha}{numeric value for render transparency. Default set to 120}
}
\description{
Render statistical map overlays on the rodent atlas
}
