% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_file_formats.R
\docType{data}
\name{rs_analysis_type_to_readable_string}
\alias{rs_analysis_type_to_readable_string}
\title{Mapping of analysis_type to readable string}
\format{
An object of class \code{list} of length 3.
}
\usage{
rs_analysis_type_to_readable_string
}
\description{
Mapping of analysis_type to readable string
}
\keyword{datasets}
