% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_file_formats.R
\docType{data}
\name{rs_atlas_files_suffix}
\alias{rs_atlas_files_suffix}
\title{List of suffixes for atlas files used in MouseSuite}
\format{
An object of class \code{list} of length 4.
}
\usage{
rs_atlas_files_suffix
}
\description{
List of suffixes for atlas files used in MouseSuite
}
\keyword{datasets}
