% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_file_formats.R
\docType{data}
\name{rs_binary_files}
\alias{rs_binary_files}
\title{List of atlas files used in MouseSuite
#' @export
rs_atlas_files <- list(
TODO: Add a list of atlas files supported by MouseSuite
)
List of binaries used in MouseSuite analysis}
\format{
An object of class \code{list} of length 2.
}
\usage{
rs_binary_files
}
\description{
List of atlas files used in MouseSuite
#' @export
rs_atlas_files <- list(
TODO: Add a list of atlas files supported by MouseSuite
)
List of binaries used in MouseSuite analysis
}
\keyword{datasets}
