% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_file_formats.R
\docType{data}
\name{rs_file_formats}
\alias{rs_file_formats}
\title{List of file formats used in MouseSuite}
\format{
An object of class \code{list} of length 8.
}
\usage{
rs_file_formats
}
\description{
List of file formats used in MouseSuite
}
\keyword{datasets}
