% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_file_formats.R
\docType{data}
\name{rs_stat_overlays}
\alias{rs_stat_overlays}
\title{List of statistical and other overlay types used in MouseSuite}
\format{
An object of class \code{list} of length 7.
}
\usage{
rs_stat_overlays
}
\description{
List of statistical and other overlay types used in MouseSuite
}
\keyword{datasets}
