% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_file_formats.R
\name{rs_volume_jacobian_file_string}
\alias{rs_volume_jacobian_file_string}
\title{Generate the designated tensor-based morphometry file}
\usage{
rs_volume_jacobian_file_string(smooth = 0)
}
\arguments{
\item{smooth}{numeric value designating the smoothing used (default is 0)}
}
\description{
Generate the designated tensor-based morphometry file
}
