% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstr.R
\docType{package}
\name{rstr}
\alias{rstr}
\title{rstr: Rodent Statistics toolbox}
\description{
The Rodent Statistics toolbox allows the application of advanced
statistical models to volumetric image, cortical surface and sulcal
curve based outputs generated from RodentSuite This enables population
or group analysis of cortical or sulcal morphology.
Available statistical models/tests are:
Pearson correlation
General linear model
ANOVA
Paired and Independent samples t-test
}
