% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstr_statops.R
\name{rstr_corr}
\alias{rstr_corr}
\title{Test for Correlation between a variable \code{corr_var} and a brain imaging phenotype.}
\usage{
rstr_corr(corr_var, rstr_data, mult_comp = "fdr")
}
\arguments{
\item{corr_var}{Character variable name. This should be present in the demographics csv file associated
with \code{rstr_data}.}

\item{rstr_data}{Object of type \code{\link{RstrData}}.}

\item{mult_comp}{method for multiple comparisons correction. The default method is "fdr". See \code{\link{rstr_p_adjust}} for valid values.}
}
\description{
Test for correlation between a brain imaging phenotype (cortical thickness, determinant
of the jacobian matrix) and \code{corr_var} using the Pearson's product moment correlation
coefficient. The brain imaging phenotype is automatically selected from the type of \code{rstr_data}.
}
\details{
\code{rstr_data} can be of the type "sba", "tbm", or "roi".
}
