% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstr_statops.R
\name{rstr_lm}
\alias{rstr_lm}
\title{Linear regression for brain imaging data.}
\usage{
rstr_lm(main_effect = "", covariates = "", rstr_data, mult_comp = "fdr")
}
\arguments{
\item{main_effect}{Character string containing an independent variable whose effect you want to measure.
It could be disease status, age, gender etc. This should strictly be a single variable. This can be
either a categorical or a continuous variable.}

\item{covariates}{Character string containing a set of other predictors (variables) in the model. If more than
one covariates are included, they should be separated by a \code{+} operator similar to an R formula.}

\item{rstr_data}{Object of type \code{\link{RstrData}}}

\item{mult_comp}{method for multiple comparisons correction. The default method is "fdr". See \code{\link{rstr_p_adjust}} for valid values.}
}
\description{
This function accepts a \code{main_effect} and a set of covariates (using the R formula notation) and performs
a linear regression including \code{main_effect + covariates}.
}
\details{
Slightly different from the standard R \code{lm} function, \code{rstr_lm} currently does not directly accept an R formula.
This could be accomodated in the future versions.
Also currently, this function returns the p-values and the t-statistics for the \code{main_effect}
only. Returning the statistics for all variables could be accomodated in the future versions.
}
