% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstr_load_roi_data.R
\name{rstr_load_roi_data}
\alias{rstr_load_roi_data}
\title{Load ROI data from subjects (BIDS compatible)}
\usage{
rstr_load_roi_data(
  subjects_dir,
  csv,
  roiids,
  roimeas,
  exclude_col,
  roistat_fileprefix,
  roilabeldescfile
)
}
\arguments{
\item{subjects_dir}{character string for subject directory.}

\item{csv}{filename of a comma separated (csv) file containing the subject demographic information.}

\item{roiids}{vector of numeric label identifiers for the region of interest (ROI) type analysis.}

\item{roimeas}{character string for the ROI measure.}

\item{exclude_col}{character string for the column in demographics csv (contains 1 or 0 for each row) specifying the subjects to exclude. 1 denotes include, 0 denotes exclude.}

\item{roistat_fileprefix}{file prefix for the ROI stats file (either .tissue.stats.txt or .roiwise.stats.txt)",}

\item{roilabeldescfile}{an xml file containing ROI label descriptions",}
}
\description{
Loads the ROI measure for statistical analysis.
}
