% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstr_statops.R
\name{rstr_ttest}
\alias{rstr_ttest}
\title{T-test for for brain imaging phenotypes}
\usage{
rstr_ttest(group_var, rstr_data, paired = FALSE, mult_comp = "fdr")
}
\arguments{
\item{group_var}{Categorical variable name. This should be present in the demographics csv file associated
with \code{rstr_data}.}

\item{rstr_data}{Object of type \code{\link{RstrData}}.}

\item{paired}{logical; is TRUE if \code{group_var} contains matching (dependent) samples. The default value is \code{FALSE}.}

\item{mult_comp}{method for multiple comparisons correction. The default method is "fdr". See \code{\link{rstr_p_adjust}} for valid values.}
}
\description{
Perform independent sample and paired sample t-tests for differences between means of brain
imaging phenotypes for a categorical variable.
}
\details{
The degrees of freedom are calculated using the Welch–Satterthwaite approximation by default.
\code{rstr_data} can be of the type "sba", "tbm", or "roi".
}
