% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstr_colormaps.R
\docType{class}
\name{RstrColormap-class}
\alias{RstrColormap-class}
\alias{RstrColormap}
\title{An S4 class for representing colormaps}
\description{
An S4 class for representing colormaps
}
\section{Slots}{

\describe{
\item{\code{cmap_type}}{A character string for the type of colormap. Valid values are "corr_values", "corr_values_masked_adjusted", "tvalues", "tvalues_adjusted", "log_pvalues", "log_pvalues_adjusted"}

\item{\code{cmap_name}}{A character string for the title of the colormap. This string will be displayed on the colorbar.}

\item{\code{values}}{A numeric vector containing the values or measures to be mapped.}

\item{\code{cex}}{Maximum absolute value of the measure to be mapped.}

\item{\code{rgbcolors}}{A matrix of RGB colors whose size is same as that of the values.}

\item{\code{lut}}{Color look up table}

\item{\code{vmin,vmax}}{Minimum and maximum values}

\item{\code{cnegmin}}{Minimum negative value}

\item{\code{cnegmax}}{Maximum negative value}

\item{\code{cposmin}}{Minimum positive value}

\item{\code{cposmax}}{Maximum positive value}
}}

