% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstroutput_class.R
\docType{class}
\name{RstrOutput-class}
\alias{RstrOutput-class}
\alias{RstrOutput}
\title{S4 class for saving results of statistical analysis}
\description{
S4 class for saving results of statistical analysis
}
\section{Slots}{

\describe{
\item{\code{outdir}}{output directory to save the results}
}}

