% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstr_rmdout_class.R
\name{RstrRmdOutput}
\alias{RstrRmdOutput}
\title{R6 super class for Rmd output functionality}
\description{
R6 super class for Rmd output functionality

R6 super class for Rmd output functionality
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{outdir}}{output directory}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RstrRmdOutput-new}{\code{RstrRmdOutput$new()}}
\item \href{#method-RstrRmdOutput-save_out}{\code{RstrRmdOutput$save_out()}}
\item \href{#method-RstrRmdOutput-finalize}{\code{RstrRmdOutput$finalize()}}
\item \href{#method-RstrRmdOutput-clone}{\code{RstrRmdOutput$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RstrRmdOutput-new"></a>}}
\if{latex}{\out{\hypertarget{method-RstrRmdOutput-new}{}}}
\subsection{Method \code{new()}}{
initialize function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RstrRmdOutput$new(outdir = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outdir}}{path to the output directory}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RstrRmdOutput-save_out"></a>}}
\if{latex}{\out{\hypertarget{method-RstrRmdOutput-save_out}{}}}
\subsection{Method \code{save_out()}}{
Save results as Rmd
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RstrRmdOutput$save_out(rstr_data, rstr_model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rstr_data}}{object of type \code{RstrData}}

\item{\code{rstr_model}}{object of type \code{RstrModel}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RstrRmdOutput-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-RstrRmdOutput-finalize}{}}}
\subsection{Method \code{finalize()}}{
finalize function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RstrRmdOutput$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RstrRmdOutput-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RstrRmdOutput-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RstrRmdOutput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
