% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstr_rmdroiout_class.R
\name{RstrRmdRoiOutput}
\alias{RstrRmdRoiOutput}
\title{R6 derived class for Rmd ROI output functionality}
\description{
R6 derived class for Rmd ROI output functionality

R6 derived class for Rmd ROI output functionality
}
\section{Super class}{
\code{\link[rstr:RstrRmdOutput]{rstr::RstrRmdOutput}} -> \code{RstrRmdRoiOutput}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RstrRmdRoiOutput-new}{\code{RstrRmdRoiOutput$new()}}
\item \href{#method-RstrRmdRoiOutput-save_out}{\code{RstrRmdRoiOutput$save_out()}}
\item \href{#method-RstrRmdRoiOutput-clone}{\code{RstrRmdRoiOutput$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rstr" data-topic="RstrRmdOutput" data-id="finalize"><a href='../../rstr/html/RstrRmdOutput.html#method-RstrRmdOutput-finalize'><code>rstr::RstrRmdOutput$finalize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RstrRmdRoiOutput-new"></a>}}
\if{latex}{\out{\hypertarget{method-RstrRmdRoiOutput-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RstrRmdRoiOutput$new(outdir = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RstrRmdRoiOutput-save_out"></a>}}
\if{latex}{\out{\hypertarget{method-RstrRmdRoiOutput-save_out}{}}}
\subsection{Method \code{save_out()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RstrRmdRoiOutput$save_out(rstr_data, rstr_model)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RstrRmdRoiOutput-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RstrRmdRoiOutput-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RstrRmdRoiOutput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
