% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstr_colormaps.R
\name{save_colorbar}
\alias{save_colorbar}
\title{Creates the display of a colorbar for an individual voxelcoordinate and overlay}
\usage{
save_colorbar(filename, lut, vmin, vmax, labeltxt)
}
\arguments{
\item{filename}{name of the file}

\item{lut}{previously generated lut file}

\item{vmin}{minimum value for the colorbar}

\item{vmax}{maximum value for the colorbar}

\item{labeltxt}{the label for the colorbar}
}
\description{
Creates the display of a colorbar for an individual voxelcoordinate and overlay
}
