% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstr_colormaps.R
\name{save_colormap_to_ini}
\alias{save_colormap_to_ini}
\title{Saves the colormap values for min and max to an ini file}
\usage{
save_colormap_to_ini(filename, rstr_cmap)
}
\arguments{
\item{filename}{name of the file}

\item{rstr_cmap}{RstrColormap object}
}
\description{
Saves the colormap values for min and max to an ini file
}
