% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstr_colormaps.R
\name{save_MouseSuiteLUT}
\alias{save_MouseSuiteLUT}
\title{Saves and writes the lut for MouseSuite use}
\usage{
save_MouseSuiteLUT(filename, lut)
}
\arguments{
\item{filename}{name of the file}

\item{lut}{previously generated lut}
}
\description{
Saves and writes the lut for MouseSuite use
}
