% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstroutput_class.R
\name{save_out}
\alias{save_out}
\alias{save_out,RstrSBAOutput-method}
\alias{save_out,RstrTBMOutput-method}
\alias{save_out,RstrDBAOutput-method}
\alias{save_out,RstrROIOutput-method}
\title{Generic save function for \code{RstrOutput}}
\usage{
save_out(rstr_out, rstr_data, rstr_model, overwrite = FALSE, ...)

\S4method{save_out}{RstrSBAOutput}(rstr_out, rstr_data, rstr_model, overwrite = F)

\S4method{save_out}{RstrTBMOutput}(rstr_out, rstr_data, rstr_model, overwrite = F, nclusters = 10)

\S4method{save_out}{RstrDBAOutput}(rstr_out, rstr_data, rstr_model, overwrite = F, nclusters = 10)

\S4method{save_out}{RstrROIOutput}(rstr_out, rstr_data, rstr_model, overwrite = F)
}
\arguments{
\item{rstr_out}{object of type \code{RstrOutput}}

\item{rstr_data}{object of type \code{RstrData}}

\item{rstr_model}{object of type \code{RstrModel}}

\item{overwrite}{logical parameter denoting if existing output directory should be overwritten or not (default is FALSE)}

\item{...}{Extra named arguments passed to save_out}

\item{nclusters}{numeric parameter denoting number of clusters (default is 10)}
}
\description{
Generic save function for \code{RstrOutput}
}
\details{
For the most part, the user will never have to call this function directly.
Instead the user should call \code{\link{save_rstr_out}}.
}
\seealso{
\code{\link{save_rstr_out}}
}
