% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstroutput_class.R
\name{save_rstr_color_files}
\alias{save_rstr_color_files}
\title{Save the color LUT and ini colormap images}
\usage{
save_rstr_color_files(
  measure,
  var_name,
  cmap_title,
  rstr_data,
  rstr_model,
  outdir
)
}
\arguments{
\item{measure}{numeric value denoting the measures used to create the color file}

\item{var_name}{string denoting name of variable that the color file is being created for}

\item{cmap_title}{string denoting the type of color map}

\item{rstr_data}{object of type \code{RstrData}}

\item{rstr_model}{object of type \code{RstrModel}}

\item{outdir}{string specifying output directory to save the results in}
}
\description{
Save the color LUT and ini colormap images
}
