% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstroutput_class.R
\name{save_rstr_out}
\alias{save_rstr_out}
\title{Save the statistical analysis output}
\usage{
save_rstr_out(
  rstr_data,
  rstr_model,
  outdir = "",
  overwrite = F,
  nclusters = 10
)
}
\arguments{
\item{rstr_data}{object of type \code{RstrData}}

\item{rstr_model}{object of type \code{RstrModel}}

\item{outdir}{output directory to save the results}

\item{overwrite}{logical parameter denoting if existing output directory should be overwritten or not (default is FALSE)}

\item{nclusters}{numeric value denoting number of clusters (default is 10)}
}
\description{
Save the statistical analysis output
}
