% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstroutput_class.R
\name{save_rstr_out_nifti_image}
\alias{save_rstr_out_nifti_image}
\title{Save the nifti image to the output file}
\usage{
save_rstr_out_nifti_image(
  measure,
  var_name,
  rstr_cmap,
  rstr_data,
  rstr_model,
  outdir
)
}
\arguments{
\item{measure}{denotes the measure used to create the output}

\item{var_name}{string denoting name of variable used by the function}

\item{rstr_cmap}{object of type \code{RstrColormap}}

\item{rstr_data}{object of type \code{RstrData}}

\item{rstr_model}{object of type \code{RstrModel}}

\item{outdir}{string specifying output directory to save the results in}
}
\description{
Save the nifti image to the output file
}
