% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstroutput_class.R
\name{save_rstr_rds}
\alias{save_rstr_rds}
\title{Save the measure to the output directory}
\usage{
save_rstr_rds(measure, var_name, label, rstr_data, rstr_model, outdir)
}
\arguments{
\item{measure}{numeric value denoting the measures used to create the output}

\item{var_name}{string denoting name of variable used by the function}

\item{label}{string denoting the label for the object}

\item{rstr_data}{object of type \code{RstrData}}

\item{rstr_model}{object of type \code{RstrModel}}

\item{outdir}{string specifying output directory to save the results in}
}
\description{
Save the measure to the output directory
}
