% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstroutput_class.R
\name{save_rstr_vol_rmd_html}
\alias{save_rstr_vol_rmd_html}
\title{Save Rstr Volumetric rmarkdown report including Rmd and the html file
This function is used to save Rmd report for both TBM and DBA}
\usage{
save_rstr_vol_rmd_html(
  outdir,
  rstr_data,
  rstr_model,
  voxelcoord,
  var_name,
  stats_string
)
}
\arguments{
\item{outdir}{string specifying output directory to save the results in}

\item{rstr_data}{object of type `RstrData`}

\item{rstr_model}{object of type `RstrModel`}

\item{voxelcoord}{list of peak voxel cluster coordinates}

\item{var_name}{string specifying the variable name}

\item{stats_string}{string specifying the statistical measure}
}
\description{
Save Rstr Volumetric rmarkdown report including Rmd and the html file
This function is used to save Rmd report for both TBM and DBA
}
