% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstr_setup.R
\name{set_brainsuite_path_in_rstr_ini}
\alias{set_brainsuite_path_in_rstr_ini}
\title{Set BrainSuite installation path in rstr.ini}
\usage{
set_brainsuite_path_in_rstr_ini(brainsuite_path)
}
\arguments{
\item{brainsuite_path}{path to the BrainSuite installation}
}
\description{
Set BrainSuite installation path in rstr.ini
}
