% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_file_formats.R
\docType{data}
\name{stats_string_to_rmd_section_title}
\alias{stats_string_to_rmd_section_title}
\title{Mapping of stats measures to readable string to be used as a section title on the Rmd report}
\format{
An object of class \code{list} of length 6.
}
\usage{
stats_string_to_rmd_section_title
}
\description{
Mapping of stats measures to readable string to be used as a section title on the Rmd report
}
\keyword{datasets}
