% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_neighbors.R
\name{add_neighbors}
\alias{add_neighbors}
\title{Add neighbors to adjacency information}
\usage{
add_neighbors(adjacency, neighs)
}
\arguments{
\item{adjacency}{Adjacency information generated by \code{spdep::poly2nb()}}

\item{neighs}{A list of regions to mark as adjacent. Uses either the index of the region or the name of the region if assigned one in \code{adjacency}}
}
\value{
A modified adjacency \code{list}.
}
\description{
Adds indices to the specified regions to indicate they should be treated as neighbors.
}
\details{
\code{add_neighbors()} is useful when adjacency information generated by \code{spdep::poly2nb()} indicates lone regions without neighbors, particularly in island counties such as the Hawaiian islands, Nantucket in Massachusetts, or San Juan in Washington. Note that \code{add_neighbors()} marks all listed counties as adjacent, so if you have a set of chaining counties where the first may not be connected to the last, several instances of \code{add_neighbors()} will be needed.
}
\examples{
ma_adj <- spdep::poly2nb(mamap)
new_neighs <- c(1, 4, 10)
ma_adj <- add_neighbors(ma_adj, new_neighs)
# age-standardize by all age groups
samples_3564 <- age_standardize(minsample, std_pop, margin = age_margin)
# age-standardize only by the first two age groups
samples_3554 <- age_standardize(minsample, std_pop[1:2], margin = age_margin, groups = 1:2)
}
