% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_standardize.R
\name{age_standardize}
\alias{age_standardize}
\title{Age-standardize samples}
\usage{
age_standardize(
  sample,
  std_pop,
  margin,
  groups = NULL,
  bind_new = FALSE,
  new_name = NULL
)
}
\arguments{
\item{sample}{Array of samples generated by \code{run_sampler()}}

\item{std_pop}{A vector of standard populations to be used for age-standardization}

\item{margin}{The margin on which the groups of interest are stratified}

\item{groups}{A vector of either indices for each group or a vector of strings for each group name. If set to \code{NULL}, will use all groups in the dataset.}

\item{bind_new}{If set to \code{TRUE}, will bind to the original sample dataset using \code{abind::abind()}. Otherwise, will generate a standalone array of samples.}

\item{new_name}{If binding the array to the original sample dataset, will assign a name to the new margin}
}
\value{
An \code{array} of age-standardized samples.
}
\description{
Consolidates a set of samples over age groups using a standard population to create age-standardized samples.
}
\examples{
std_pop <- c(113154, 100640, 95799)
age_margin <- 2
# age-standardize by all age groups
samples_3564 <- age_standardize(minsample, std_pop, margin = age_margin)
# age-standardize only by the first two age groups
samples_3554 <- age_standardize(minsample, std_pop[1:2], margin = age_margin, groups = 1:2)
# bind age-standardized samples to original samples
samples_as <- age_standardize(
  minsample,
  std_pop,
  margin = age_margin,
  bind_new = TRUE,
  new_name = "35-64"
)
}
