% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_standardize.R
\name{aggregate_pop}
\alias{aggregate_pop}
\title{Aggregate population arrays}
\usage{
aggregate_pop(pop, margin)
}
\arguments{
\item{pop}{Population array}

\item{margin}{The margin on which the groups of interest are stratified}
}
\description{
Aggregate population arrays
}
