% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_standardize.R
\name{load_pop}
\alias{load_pop}
\title{Load population array}
\usage{
load_pop(name, dir = tempdir())
}
\value{
The population array for the specified model.
}
\description{
Imports the population array from the model directory.
}
\examples{
initialize_model("test", tempdir(), miheart, miadj, show_plots = FALSE)
pop <- load_pop("test", tempdir())
\dontshow{
unlink(paste0(tempdir(), "\\\\test"), recursive = TRUE)
}
}
