\name{RSTr-package}
\alias{RSTr-package}
\alias{RSTr}
\docType{package}
\title{
\packageTitle{RSTr}
}
\description{
\packageDescription{RSTr}
}
\details{
The RSTr package is designed to spatially smooth discrete small-area event rates using information from neighboring spatial regions. See `browseVignettes("RSTr")` for a series of tutorials on basic usage of the RSTr functions.
}
\author{
\packageAuthor{RSTr}

Maintainer: \packageMaintainer{RSTr}
}
\references{
Besag, J., York, J. and Mollie, A. (1991) Bayesian Image Restoration with Two Applications in Spatial Statistics (with Discussion). Annals of the Institute of Statistical Mathematics, 43, 1-59. https://doi.org/10.1007/BF00116466
Multivariate spatiotemporal modeling of age-specific stroke mortality. (2017). Ann. Appl. Stat., 11(4), 2165–2177. https://doi.org/10.1214/17-AOAS1068
Song, G. (n.d.). Estimating the Informativeness of the Conditional Autoregressive Model Framework with Applications. Drexel University Libraries. https://doi.org/10.17918/00001456
}
\keyword{package}
