\name{RSTr-package}
\alias{RSTr-package}
\alias{RSTr}
\docType{package}
\title{
\packageTitle{RSTr}
}
\description{
\packageDescription{RSTr}
}
\details{
The RSTr package uses Bayesian spatiotemporal modeling to spatially smooths discrete small-area event rates using information from neighboring spatial regions. See `browseVignettes("RSTr")` for a series of tutorials on basic usage of the RSTr functions.
}
\author{
\packageAuthor{RSTr}

Maintainer: \packageMaintainer{RSTr}
}
\references{

Besag, J., York, J., and Mollié, A. (1991).
Bayesian Image Restoration with Two Applications in Spatial Statistics (with Discussion).
*Annals of the Institute of Statistical Mathematics*, **43**, 1–59.
\doi{10.1007/BF00116466}

Gelfand, A. E., & Vounatsou, P. (2003).
Proper multivariate conditional autoregressive models for spatial data analysis.
*Biostatistics*, **4**(1), 11–25.
\doi{10.1093/biostatistics/4.1.11}

Quick, et al. (2017).
Multivariate spatiotemporal modeling of age-specific stroke mortality.
*Annals of Applied Statistics*, **11**(4), 2165–2177.
\doi{10.1214/17-AOAS1068}

Quick, et al. (2021).
Evaluating the informativeness of the Besag-York-Mollié CAR model.
*Spatial and Spatio-temporal Epidemiology*, **37**, 100420.
\doi{10.1016/j.sste.2021.100420}

}

\keyword{package}
