% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sampler.R
\name{run_sampler}
\alias{run_sampler}
\title{Run Gibbs sampler}
\usage{
run_sampler(
  name,
  dir = getwd(),
  iterations = 6000,
  .show_plots = TRUE,
  .discard_burnin = FALSE
)
}
\arguments{
\item{name}{Name of model and corresponding folder}

\item{dir}{Directory where model lives}

\item{.show_plots}{Show or hide traceplots as}

\item{.discard_burnin}{If set to \code{TRUE}, won't save burn-in samples}
}
\description{
Run Gibbs sampler
}
