% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sampler.R
\name{run_sampler}
\alias{run_sampler}
\title{Run Gibbs sampler}
\usage{
run_sampler(
  name,
  dir = tempdir(),
  iterations = 6000,
  .show_plots = TRUE,
  .show_progress = TRUE,
  .discard_burnin = FALSE
)
}
\arguments{
\item{name}{Name of model and corresponding folder}

\item{dir}{Directory where model lives}

\item{iterations}{Specifies number of iterations to run}

\item{.show_plots}{If set to \code{FALSE}, hides traceplots}

\item{.show_progress}{If set to \code{FALSE}, hides progress bar}

\item{.discard_burnin}{If set to \code{TRUE}, won't save burn-in samples}
}
\value{
No output, saves sampler output to \code{dir}
}
\description{
\code{run_sampler()} generates samples for model \code{name} in \code{dir}. The model used to generate samples (e.g., MSTCAR, MCAR, UCAR) along with the model's other parameters are specified in \code{initialize_model()}.
}
\examples{
data_min = lapply(miheart, \(x) x[1:2, 1:3, 1:3])
adj_min = list(2, 1)
initialize_model("test", tempdir(), data_min, adj_min)
run_sampler("test", .show_plots = FALSE)

# delete temp model files generated by example
unlink(paste0(tempdir(), "\\\\test"), recursive = TRUE)
}
