% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_suppress_estimates.R
\name{suppress_estimates}
\alias{suppress_estimates}
\title{Suppress estimates based on reliability criteria}
\usage{
suppress_estimates(RSTr_obj, pop_threshold = Inf)
}
\arguments{
\item{RSTr_obj}{An \code{RSTr} model object.}

\item{rates_per}{The desired scaling for estimate rates.}

\item{standardized}{If \code{RSTr_obj} contains age-standardized rates, shows the age-standardized rates. If set to \code{FALSE}, always shows the non-age-standardized rates.}
}
\value{
A long \code{table} containing region/group/time period names, estimates, credible intervals, relative precisions, and the associated event/population counts.
}
\description{
Generates suppressed estimates for an \code{RSTr} model object with a given relative precision and population threshold.
}
\details{
While the \code{pop_threshold} argument is optional, population thresholds are necessary for non-enhanced models. Population thresholds should only be omitted for enhanced CAR models, such as the EUCAR.
}
\examples{
std_pop <- c(113154, 100640, 95799)
data_min <- lapply(miheart, \(x) x[1:2, 1:3, 1:3])
adj_min <- list(2, 1)
mod_mst <- mstcar("test", data_min, adj_min, tempdir(), show_plots = FALSE)
estimates_table <- get_estimates(mod_mst)
mod_mst <- age_standardize(mod_mst, std_pop, "35-64")
estimates_table_as <- get_estimates(mod_mst)
\dontshow{
unlink(paste0(tempdir(), "\\\\test"), recursive = TRUE)
}
}
