% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_model_auxiliary.R
\name{update_model}
\alias{update_model}
\title{\code{update_model()} generates further samples for model \code{name} in \code{dir}. The model used to generate samples (e.g., MSTCAR, MCAR, UCAR) along with the model's other parameters are specified in \code{*car()}.}
\usage{
update_model(RSTr_obj, iterations = 6000, show_plots = TRUE, verbose = TRUE)
}
\arguments{
\item{RSTr_obj}{The \code{RSTr} model object to generate samples for}

\item{iterations}{Specifies number of iterations to run}

\item{show_plots}{If set to \code{FALSE}, hides traceplots}

\item{verbose}{If set to \code{FALSE}, hides progress bar and other messages}
}
\value{
No output, saves sampler output to \code{dir}
}
\description{
\code{update_model()} generates further samples for model \code{name} in \code{dir}. The model used to generate samples (e.g., MSTCAR, MCAR, UCAR) along with the model's other parameters are specified in \code{*car()}.
}
\examples{
data_min <- lapply(miheart, \(x) x[1:2, 1:3, 1:3])
adj_min <- list(2, 1)
mod_mst <- mstcar("mod_mst", data_min, adj_min, tempdir())
mod_mst <- update_model(mod_mst, show_plots = FALSE, verbose = FALSE)
\dontshow{
unlink(paste0(tempdir(), "\\\\mod_mst"), recursive = TRUE)
}
}
